# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech Challenge API",
        "description": "Documentação das APIs de autoatendimento da lanchonete.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os produtos",
                "description": "Retorna uma lista de todos os produtos cadastrados no sistema.",
                "operationId": "6d8ae74dcafe19c4813904d4ea9e4c57",
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Produto A"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Descrição do produto A."
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 29.99
                                            },
                                            "discount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5
                                            },
                                            "stockQuantity": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "category": {
                                                "type": "integer",
                                                "example": 2,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Cadastrar um novo produto",
                "description": "Cria um novo produto no sistema com as informações fornecidas.",
                "operationId": "9f0b814cd8d69075ed70d4cc79c9f417",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description",
                                    "stockQuantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Produto A"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição detalhada do produto A."
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://meusite.com/imagem.jpg"
                                    },
                                    "stockQuantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "category": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto cadastrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação: dados inválidos ou produto já cadastrado."
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar um produto existente",
                "description": "Atualiza os dados de um produto existente no sistema com base no ID fornecido.",
                "operationId": "1b73c420e0cf9fb1cb88dcf445b9428b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do produto a ser atualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Produto B"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do produto atualizado."
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully"
                                        },
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Produto B"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 49.99
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descrição detalhada do produto atualizado."
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação: dados inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "O campo nome é obrigatório."
                                                ],
                                                "price": [
                                                    "O preço deve ser um número positivo."
                                                ]
                                            },
                                            "additionalProperties": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "property": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}